+++ app/components/Transfers/DisplayRegisterData.tsx
@@ -1,0 +1,34 @@
+import React from 'react';
+import { RegisterData } from '~/utils/types';
+import { useAccountName } from '~/utils/dataHooks';
+import DisplayTransactionExpiryTime from '../DisplayTransactionExpiryTime';
+import { dateFromTimeStamp } from '~/utils/timeHelpers';
+import DisplayAddress from '../DisplayAddress';
+
+import styles from './transferDetails.module.scss';
+
+interface Props {
+    transaction: RegisterData;
+}
+
+/**
+ * Displays an overview of register data transaction.
+ */
+export default function DisplayRegisterData({ transaction }: Props) {
+    const senderName = useAccountName(transaction.sender);
+    return (
+        <>
+            <p className={styles.title}>From Account:</p>
+            <p className={styles.name}>{senderName}</p>
+            <DisplayAddress
+                address={transaction.sender}
+                lineClassName={styles.address}
+            />
+            <h5 className={styles.title}>Registered data:</h5>
+            <p className={styles.memo}>{transaction.payload.data}</p>
+            <DisplayTransactionExpiryTime
+                expiryTime={dateFromTimeStamp(transaction.expiry)}
+            />
+        </>
+    );
+}