+++ app/pages/multisig/AccountTransactions/OldBakerFlows/RemoveBaker.tsx
@@ -30,14 +30,14 @@ import { getNextAccountNonce } from '~/node/nodeRequests';
 import errorMessages from '~/constants/errorMessages.json';
 import LoadingComponent from '../LoadingComponent';
 import {
-    BakerSubRoutes,
+    AccountTransactionSubRoutes,
     getLocationAfterAccounts,
 } from '~/utils/accountRouterHelpers';
 import DatePicker from '~/components/Form/DatePicker';
 import { isMultiSig } from '~/utils/accountHelpers';
 import { findAccountTransactionHandler } from '~/utils/transactionHandlers/HandlerFinder';
 
-import styles from '../MultisignatureAccountTransactions.module.scss';
+import styles from '../../common/MultiSignatureFlowPage.module.scss';
 
 interface PageProps {
     exchangeRate: Fraction;
@@ -123,7 +123,7 @@ function RemoveBakerPage({ exchangeRate }: PageProps) {
                             className={styles.stretchColumn}
                         >
                             <div className={styles.columnContent}>
-                                <div className={styles.flex1}>
+                                <div className="flexChildFill">
                                     <PickAccount
                                         setAccount={setAccount}
                                         chosenAccount={account}
@@ -165,13 +165,15 @@ function RemoveBakerPage({ exchangeRate }: PageProps) {
                             </div>
                         </Columns.Column>
                     </Route>
-                    <Route path={`${path}/${BakerSubRoutes.expiry}`}>
+                    <Route
+                        path={`${path}/${AccountTransactionSubRoutes.expiry}`}
+                    >
                         <Columns.Column
                             header="Transaction expiry time"
                             className={styles.stretchColumn}
                         >
                             <div className={styles.columnContent}>
-                                <div className={styles.flex1}>
+                                <div className="flexChildFill">
                                     <p className="mT0">
                                         Choose the expiry date for the
                                         transaction.
@@ -217,7 +219,7 @@ function RemoveBakerPage({ exchangeRate }: PageProps) {
                                             .then(() =>
                                                 dispatch(
                                                     push(
-                                                        `${url}/${BakerSubRoutes.sign}`
+                                                        `${url}/${AccountTransactionSubRoutes.sign}`
                                                     )
                                                 )
                                             )
@@ -233,7 +235,7 @@ function RemoveBakerPage({ exchangeRate }: PageProps) {
                             </div>
                         </Columns.Column>
                     </Route>
-                    <Route path={`${path}/${BakerSubRoutes.sign}`}>
+                    <Route path={`${path}/${AccountTransactionSubRoutes.sign}`}>
                         <Columns.Column
                             header="Signature and hardware wallet"
                             className={styles.stretchColumn}